LC-find-substring-with-given-hash-value [Runtime: 2417 ms; Memory: 325.1 MB]

+++ string/find-substring-with-given-hash-value.cpp
@@ -11,7 +11,7 @@ public:
       for(int i=0;i<s.size();i++){
         string subs=s.substr(i,k);
         int hashval=0;
-        int p=1;
+        long long p=1;
           for(int j=0;j<subs.size();j++){
 
             hashval=(hashval+hv(subs[j])*p)%modulo;
